JBPM-7760 - Provide Test Helper Classes for Case Management and jBPM services
