+++ jbpm-case-mgmt/jbpm-case-mgmt-impl/src/main/resources/META-INF/services/org.jbpm.casemgmt.api.utils.CaseServiceConfigurator
@@ -1,0 +1,1 @@
+org.jbpm.casemgmt.impl.utils.DefaultCaseServiceConfigurator
