+++ jbpm-services/jbpm-kie-services/src/main/resources/META-INF/services/org.jbpm.services.api.utils.KieServiceConfigurator
@@ -1,0 +1,1 @@
+org.jbpm.kie.services.impl.utils.DefaultKieServiceConfigurator
