+++ jbpm-services/jbpm-services-cdi/src/test/java/org/jbpm/services/cdi/test/UserTaskServiceCDIImplTest.java
@@ -24,7 +24,6 @@ import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ArchivePaths;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.jbpm.kie.services.test.TestIdentityProvider;
 import org.jbpm.kie.services.test.UserTaskServiceImplTest;
 import org.jbpm.services.api.DefinitionService;
 import org.jbpm.services.api.DeploymentService;
@@ -33,10 +32,9 @@ import org.jbpm.services.api.RuntimeDataService;
 import org.jbpm.services.api.UserTaskService;
 import org.jbpm.shared.services.impl.TransactionalCommandService;
 import org.jbpm.shared.services.impl.commands.UpdateStringCommand;
+import org.jbpm.test.services.TestIdentityProvider;
 import org.junit.After;
-import org.junit.Before;
 import org.junit.runner.RunWith;
-import org.kie.internal.identity.IdentityProvider;
 
 @RunWith(Arquillian.class)
 public class UserTaskServiceCDIImplTest extends UserTaskServiceImplTest {
