+++ jbpm-services/jbpm-services-ejb/jbpm-services-ejb-client/pom.xml
@@ -43,6 +43,11 @@
     </dependency>
     <dependency>
       <groupId>org.jbpm</groupId>
+      <artifactId>jbpm-test</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.jbpm</groupId>
       <artifactId>jbpm-kie-services</artifactId>
       <type>test-jar</type>
       <scope>test</scope>
